authorization fixes
