+++ src/components/Home.js
@@ -15,27 +15,27 @@ const validCodes = {
 
 
 const Home = () => {
-  // **SECTION** guest validation
-  // to be hooked up to a database eventually 
-  const [code, setCode] = useState(0)
-  const [name, setName] = useState('');
+  // // **SECTION** guest validation
+  // // to be hooked up to a database eventually 
+  // const [code, setCode] = useState(0)
+  // const [name, setName] = useState('');
 
-  // gets the code out of the textbox
-  const handleInputChange = (event) => {
-    setCode(event.target.value);
-  };
+  // // gets the code out of the textbox
+  // const handleInputChange = (event) => {
+  //   setCode(event.target.value);
+  // };
 
-  // sets the user code as the State
-  // feature desire: the code will query a db where there will be a guest-code mapping 
-  // from there, the user will be able to see information specific to their attendance
-  // eg: wedding party will see rehearsal dinner, morning of information
-  const handleSubmit = () => {
-    if (validCodes[code]) {
-      setName(validCodes[code]);
-    } else {
-      alert('Invalid code. Please try again.');
-    }
-  };
+  // // sets the user code as the State
+  // // feature desire: the code will query a db where there will be a guest-code mapping 
+  // // from there, the user will be able to see information specific to their attendance
+  // // eg: wedding party will see rehearsal dinner, morning of information
+  // const handleSubmit = () => {
+  //   if (validCodes[code]) {
+  //     setName(validCodes[code]);
+  //   } else {
+  //     alert('Invalid code. Please try again.');
+  //   }
+  // };
   
 
   // **SECTION** Date Stuff
@@ -58,42 +58,14 @@ const Home = () => {
 
   return (
     <Container>
-      {/* 
-      TODO: this needs to be moved to <App>.
-      Temporary: if else handles user access controls.
-      If there is any of the above codes (1-8), it'll let you in. 
-      Eventually these codes will be hashes (?) where the user will enter the code
-      and a db will be queried getting specific information and generating content 
-      specific to that user
-      */}
-      {/* if no current name is in the state, show this code-entry box */}
-      {!name ? (
-        <Box>
-          <Typography variant="h4" gutterBottom>
-            Enter Your Code
-          </Typography>
-          <Box  sx={{display:'flex', p:2, gap: 2}}>
-            <TextField
-              label="Invite Code"
-              variant="outlined"
-              value={code}
-              onChange={handleInputChange}
-              />
-            <Button variant="contained" color="primary" onClick={handleSubmit}>
-              Submit
-            </Button>
-          </Box>
-        </Box>
-      ) :  
-      // if there is a name, show the content
-      (<Grid container spacing={1}>
+      <Grid container spacing={1}>
         <Grid item xs={12} md={6}>
           <Box sx={{
             p: 1,
             gap: 1
           }}>
             <Typography variant='h2'>
-              Welcome, {name}
+              Welcome
             </Typography>
             <Typography variant="h1"> 
               The Marriage of Isabel Haziomeric & Daniel Meleras
@@ -118,7 +90,6 @@ const Home = () => {
             />
         </Grid>
       </Grid>
-      )}
     </Container>
   );
 };
